requirements.txt
azure-identity==1.12.0
azure-storage-blob==12.14.1
azure-mgmt-storage==21.0.0
azure-keyvault-secrets==4.6.0 